+++ src/Button.jsx
@@ -1,9 +1,7 @@
-import React from 'react';
-
-export default function Button({ children, onClick }) {
+export default function Button({ value, onClick }) {
   return (
-    <button type="button" onClick={() => onClick(children)}>
-      {children}
+    <button type="button" onClick={() => onClick(value)}>
+      {value}
     </button>
   );
 }
