+++ src/Counter.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function Counter({ count, onClick }) {
   return (
     <button type="button" onClick={onClick}>
